convite administrador

+++ server/service/userService.js
@@ -189,6 +189,29 @@ const findByUnidade = async (unidadeId) => {
     return promise;
 }
 
+const findAdministradores = async () => {
+    const promise = new Promise( (resolve, reject) => {
+        var MongoClient = require( 'mongodb' ).MongoClient;
+        MongoClient.connect( process.env.MONGO_URIS, { useUnifiedTopology: false }, function( err, client ) {
+            if(err) return reject(err);
+            const db = client.db(dbName);
+            
+            const collection = db.collection(collectionName);
+
+            collection.find({ role: 'ADMINISTRADOR' }, { projection: { password: 0 } }).toArray(function(err, result) {
+                if(err) {
+                    reject(err);
+                } else {
+                    resolve(result);
+                }
+            });
+        });
+
+    });
+
+    return promise;
+}
+
 const validateResetToken = async (id, token) => {
     const promise = new Promise( (resolve, reject) => {
         var MongoClient = require( 'mongodb' ).MongoClient;
@@ -235,4 +258,4 @@ const validateInvitationToken = async (id, token) => {
     return promise;
 }
 
-module.exports = {  findAll, deleteAll, insertAll, replaceOne, insertOne, findByEmail, findById, findByUnidade, validateResetToken, validateInvitationToken };
+module.exports = {  findAll, deleteAll, insertAll, replaceOne, insertOne, findByEmail, findById, findByUnidade, validateResetToken, validateInvitationToken, findAdministradores };
